refactoring supportconfig
