+++ mgrpxy/cmd/support/config/extractor.go
@@ -8,7 +8,9 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
+	"github.com/uyuni-project/uyuni-tools/shared/kubernetes"
 	. "github.com/uyuni-project/uyuni-tools/shared/l10n"
+	"github.com/uyuni-project/uyuni-tools/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
 	"github.com/uyuni-project/uyuni-tools/shared/utils"
 )
@@ -20,35 +22,22 @@ func extract(globalFlags *types.GlobalFlags, flags *configFlags, cmd *cobra.Comm
 		return utils.Errorf(err, L("failed to create temporary directory"))
 	}
 	defer os.RemoveAll(tmpDir)
+	var fileList []string
+	if podman.HasService(podman.ServerService) {
+		fileList, err = podman.RunSupportConfigOnHost(tmpDir)
+	}
 
-	hostSupportConfigFiles, err := utils.RunSupportConfigOnHost(tmpDir)
+	if utils.IsInstalled("kubectl") && utils.IsInstalled("helm") {
+		fileList, err = kubernetes.RunSupportConfigOnHost(tmpDir)
+	}
+
 	if err != nil {
 		return err
 	}
 
-	// TODO Get cluster infos in case of kubernetes
-
-	if err := utils.CreateSupportConfigTarball(flags.Output, hostSupportConfigFiles); err != nil {
+	if err := utils.CreateSupportConfigTarball(flags.Output, fileList); err != nil {
 		return err
 	}
 
 	return nil
 }
-
-//func getAllProxyCnx(backend string) map[*shared.Connection]bool {
-//	cnxs := make(map[*shared.Connection]bool)
-//
-//	/* this is as hack but it works. We loop for podman proxy container name since the proxy filter for kubernetes
-//	* is just one. Storing the context in a unique list, we would have one result if kubernetes and one for each
-//	* container for podman.
-//	 */
-//	for _, container := range podman.ProxyContainerNames {
-//		cnx := shared.NewConnection(backend, container, kubernetes.ProxyFilter)
-//		if cnxs[cnx] {
-//			continue
-//		} else {
-//			cnxs[cnx] = true
-//		}
-//	}
-//	return cnxs
-//}
